Checkpoint

+++ scripts/extract_facenet_single.py
@@ -8,6 +8,8 @@ import time
 from multiprocessing import Process
 from multiprocessing.pool import Pool, ThreadPool
 
+import ffmpeg
+import numpy as np
 import torch
 import torch.backends.cudnn as cudnn
 import torch.nn.parallel
@@ -18,7 +20,7 @@ from tqdm import tqdm
 
 import pretorched
 from data import VideoFolder, VideoZipFile, video_collate_fn
-from models import FaceModel, deepmmag
+from models import FaceModel
 from pretorched.runners.utils import AverageMeter, ProgressMeter
 from pretorched.utils import str2bool
 
@@ -28,27 +30,35 @@ try:
 except ModuleNotFoundError:
     raise ModuleNotFoundError
 
-DEEPFAKE_DATA_ROOT = os.path.join(os.environ['DATA_ROOT'], 'DeepfakeDetection')
-
 
 def parse_args():
     parser = argparse.ArgumentParser(description='Face Extraction')
-    parser.add_argument('--filename', type=str, default='')
+    parser.add_argument('--dataset', type=str, default='DeepfakeDetection')
+    parser.add_argument('--filename', type=str, default=None)
     parser.add_argument('--part', type=str, default='dfdc_train_part_0')
     parser.add_argument('--magnify_motion', default=False, type=str2bool)
     parser.add_argument('--overwrite', default=False, type=str2bool)
     parser.add_argument('--remove_frames', default=True, type=str2bool)
-    parser.add_argument('--use_zip', default=True, type=str2bool)
+    parser.add_argument('--use_zip', default=False, type=str2bool)
     parser.add_argument('--video_rootdir', default='videos', type=str)
     parser.add_argument('--face_rootdir', default='facenet_videos', type=str)
     parser.add_argument('--chunk_size', default=150, type=int)
+    parser.add_argument('--num_workers', default=2, type=int)
     args = parser.parse_args()
-    args.video_dir = os.path.join(DEEPFAKE_DATA_ROOT, args.video_rootdir, args.part)
-    args.face_dir = os.path.join(DEEPFAKE_DATA_ROOT, args.face_rootdir, args.part)
+    DEEPFAKE_DATA_ROOT = os.path.join(os.environ['DATA_ROOT'], args.dataset)
+    if args.dataset == 'DeepfakeDetection':
+        args.video_dir = os.path.join(DEEPFAKE_DATA_ROOT, args.video_rootdir, args.part)
+        args.face_dir = os.path.join(DEEPFAKE_DATA_ROOT, args.face_rootdir, args.part)
+    elif args.dataset == 'FaceForensics':
+        args.video_dir = os.path.join(DEEPFAKE_DATA_ROOT, args.part, args.video_rootdir)
+        args.face_dir = os.path.join(DEEPFAKE_DATA_ROOT, args.part, args.face_rootdir)
+    else:
+        args.video_dir = os.path.join(DEEPFAKE_DATA_ROOT, args.video_rootdir, args.part)
+        args.face_dir = os.path.join(DEEPFAKE_DATA_ROOT, args.face_rootdir, args.part)
     return args
 
 
-def main(filename, video_dir, face_dir, size=360, margin=100, fdir_tmpl='face_{}', tmpl='{:06d}.jpg',
+def main(video_dir, face_dir, filename=None, size=360, margin=100, fdir_tmpl='face_{}', tmpl='{:06d}.jpg',
          metadata_fname='face_metadata.json', step=1, batch_size=1, chunk_size=300, num_workers=2,
          overwrite=False, remove_frames=True, magnify_motion=False, use_zip=True, **kwargs):
 
@@ -62,8 +72,12 @@ def main(filename, video_dir, face_dir, size=360, margin=100, fdir_tmpl='face_{}
     else:
         dataset = VideoFolder(video_dir, step=step)
 
-    print(f'Processing: {filename} in {video_dir}')
-    dataset.videos_filenames = [filename]
+    print(f'Processing: {video_dir}')
+    if not overwrite:
+        dataset.videos_filenames = filter_filenames(dataset.videos_filenames, face_dir)
+
+    if filename is not None:
+        dataset.videos_filenames = [filename]
 
     dataloader = DataLoader(dataset, batch_size=batch_size,
                             shuffle=False, num_workers=num_workers,
@@ -79,10 +93,15 @@ def main(filename, video_dir, face_dir, size=360, margin=100, fdir_tmpl='face_{}
                       select_largest=False,
                       chunk_size=chunk_size)
 
-    run_motion_mag = deepmmag.get_motion_mag() if magnify_motion else None
+    if magnify_motion:
+        from models import deepmmag
+        run_motion_mag = deepmmag.get_motion_mag()
+    else:
+        run_motion_mag = None
 
-    batch_time = AverageMeter('Time', ':6.3f')
-    progress = ProgressMeter(len(dataset), [batch_time], prefix='Facenet Extraction and MM: ')
+    # batch_time = AverageMeter('Time', ':6.3f')
+    # data_time = AverageMeter('Time', ':6.3f')
+    # progress = ProgressMeter(len(dataset), [batch_time, data_time], prefix='Facenet Extraction and MM: ')
 
     # switch to evaluate mode
     model.eval()
@@ -93,10 +112,11 @@ def main(filename, video_dir, face_dir, size=360, margin=100, fdir_tmpl='face_{}
             with contextlib.suppress(RuntimeWarning):
 
                 filenames, x, _ = next(dataloader)
-                print(f'Extracting faces from: {filenames}')
-
-                face_images = model.get_faces(x)
-                torch.cuda.empty_cache()
+                # for f in filenames:
+                # if os.path.exists(os.path.join(face_dir, os.path.basename(f))):
+                # print(f'Skipping {f}')
+                # continue
+                face_images = model.get_faces(x, to_pil=magnify_motion)
 
                 for filename, face_images in zip(filenames, face_images):
                     save_dir = os.path.join(face_dir, os.path.basename(filename))
@@ -106,10 +126,10 @@ def main(filename, video_dir, face_dir, size=360, margin=100, fdir_tmpl='face_{}
                                    remove_frames=remove_frames)
 
                 # measure elapsed time
-                batch_time.update(time.time() - end)
+                # batch_time.update(time.time() - end)
                 end = time.time()
 
-                progress.display(i)
+                # progress.display(i)
 
 
 def save_face_data(save_dir, face_images, run_motion_mag=None, size=360, margin=100, fdir_tmpl='face_{}',
@@ -134,20 +154,53 @@ def save_face_data(save_dir, face_images, run_motion_mag=None, size=360, margin=
         json.dump(metadata, f)
 
     for face_num, faces in face_images.items():
-        num_images = len(faces)
-        out_filename = os.path.join(save_dir, fdir_tmpl.format(face_num), tmpl)
-        names = (out_filename.format(i) for i in range(1, num_images + 1))
-        save_images(faces, names, num_workers=num_images)
-
         video_path = os.path.join(save_dir, fdir_tmpl.format(face_num))
-        frames_to_video(video_path)
-        # motion_magnification(video_path, remove_frames)
-        if run_motion_mag is not None:
-            mm_out_dir = run_motion_mag(video=video_path, output=video_path + '_mm')
-        if remove_frames:
-            os.system(f'rm -rf {video_path}')
+        if run_motion_mag is None:
+            array_to_video(faces, video_path + '.mp4')
+        else:
+            num_images = len(faces)
+            out_filename = os.path.join(save_dir, fdir_tmpl.format(face_num), tmpl)
+            names = (out_filename.format(i) for i in range(1, num_images + 1))
+            save_images(faces, names, num_workers=num_images)
+
+            video_path = os.path.join(save_dir, fdir_tmpl.format(face_num))
+            frames_to_video(video_path)
+            # motion_magnification(video_path, remove_frames)
             if run_motion_mag is not None:
-                os.system(f'rm -rf {mm_out_dir}')
+                mm_out_dir = run_motion_mag(video=video_path, output=video_path + '_mm')
+            if remove_frames:
+                os.system(f'rm -rf {video_path}')
+                if run_motion_mag is not None:
+                    os.system(f'rm -rf {mm_out_dir}')
+
+
+def vidwrite(images, filename, framerate=30, vcodec='libx264'):
+    if not isinstance(images, np.ndarray):
+        images = np.asarray(images)
+    n, height, width, channels = images.shape
+    process = (
+        ffmpeg
+        .input('pipe:', format='rawvideo', pix_fmt='rgb24', s='{}x{}'.format(width, height))
+        .output(filename, pix_fmt='yuv420p', vcodec=vcodec, r=framerate)
+        .global_args('-loglevel', 'error')
+        .overwrite_output()
+        .run_async(pipe_stdin=True)
+    )
+    for frame in images:
+        process.stdin.write(
+            frame
+            .astype(np.uint8)
+            .tobytes()
+        )
+    process.stdin.close()
+    process.wait()
+
+
+def array_to_video(images, filename):
+    p = Process(target=vidwrite,
+                args=(images, filename),
+                kwargs={'vcodec': DEFAULT_VIDEO_CODEC})
+    p.start()
 
 
 def frames_to_video(video_path):
@@ -159,6 +212,7 @@ def frames_to_video(video_path):
 
 
 def motion_magnification(video_path, remove_frames):
+    from models import deepmmag
     p = Process(target=deepmmag.motion_magnify,
                 kwargs={'video': video_path,
                         'output': video_path + '_mm',
@@ -168,6 +222,7 @@ def motion_magnification(video_path, remove_frames):
 
 
 def _motion_magnification(video_path, remove_frames):
+    from models import deepmmag
     with Pool(2) as pool:
         res = pool.apply_async(deepmmag.motion_magnify,
                                kwds={'video': video_path,
@@ -196,13 +251,22 @@ def save_image(args):
         pass
 
 
-def filter_filenames(video_filenames, face_dir):
+def filter_filenames(video_filenames, face_root):
     filtered_filename = []
     for f in video_filenames:
-        frame_dir = os.path.join(face_dir, os.path.basename(f))
-        if os.path.exists(frame_dir):
-            if os.listdir(frame_dir):
-                print(f'Skipping {frame_dir}')
+        face_dir = os.path.join(face_root, os.path.basename(f))
+        if os.path.exists(face_dir):
+            fm = os.path.join(face_dir, 'face_metadata.json')
+            try:
+                with open(fm) as jf:
+                    metadata = json.load(jf)
+                for face_name in metadata['face_names']:
+                    if not os.path.exists(os.path.join(face_dir, face_name + '.mp4')):
+                        raise FileNotFoundError
+            except FileNotFoundError:
+                pass
+            else:
+                print(f'Skipping {face_dir}')
                 continue
         filtered_filename.append(f)
     return filtered_filename
